+++ Empresa.hpp
@@ -30,7 +30,7 @@ class Empresa {
   public:
     static Empresa *instEmpresa();
     Funcionario* adicionaFuncionario(TipoPessoa tipo, std::string cadastro, std::string nome,
-                              std::string email, std::pair<int, int> endereco, unsigned anoNasc, unsigned mesNasc, unsigned diaNasc,
+                              std::string email, std::pair<float, float> endereco, unsigned anoNasc, unsigned mesNasc, unsigned diaNasc,
                                 Departamento* departamento, Cargo *cargo, unsigned anoCria, unsigned mesCria, unsigned diaCria, float salario);
     Cliente* adicionarCliente(std::string telefone, std::string nome, std::string cadastro,
           std::string email, enum TipoPessoa tipo);
